# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FlawMind API Documentation",
        "description": "API documentation for FlawMind ERP system",
        "contact": {
            "name": "API Support",
            "email": "support@flawmind.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/v1/business/register": {
            "post": {
                "tags": [
                    "Business - Settings"
                ],
                "summary": "Register a new business",
                "description": "Create a new business with an owner account and a first location. Returns a bearer token for immediate use. This endpoint is public and does not require authentication.",
                "operationId": "f484d18094db18ee17f64ae9336e7f64",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "currency_id",
                                    "time_zone",
                                    "fy_start_month",
                                    "accounting_method",
                                    "first_name",
                                    "username",
                                    "password",
                                    "location_name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Business name",
                                        "type": "string",
                                        "example": "My Business"
                                    },
                                    "currency_id": {
                                        "description": "Currency ID from currencies table",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "time_zone": {
                                        "description": "PHP timezone string",
                                        "type": "string",
                                        "example": "Asia/Dhaka"
                                    },
                                    "fy_start_month": {
                                        "description": "Fiscal year start month (1-12)",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "accounting_method": {
                                        "type": "string",
                                        "enum": [
                                            "fifo",
                                            "lifo",
                                            "avco"
                                        ],
                                        "example": "fifo"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-01-01",
                                        "nullable": true
                                    },
                                    "tax_label_1": {
                                        "type": "string",
                                        "example": "VAT",
                                        "nullable": true
                                    },
                                    "tax_number_1": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "tax_label_2": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "tax_number_2": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "surname": {
                                        "type": "string",
                                        "example": "Mr.",
                                        "nullable": true
                                    },
                                    "first_name": {
                                        "description": "Owner first name",
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Doe",
                                        "nullable": true
                                    },
                                    "username": {
                                        "description": "Owner username (min 4 chars, unique)",
                                        "type": "string",
                                        "example": "johndoe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com",
                                        "nullable": true
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    },
                                    "language": {
                                        "type": "string",
                                        "example": "en",
                                        "nullable": true
                                    },
                                    "location_name": {
                                        "description": "First location name",
                                        "type": "string",
                                        "example": "Head Office"
                                    },
                                    "landmark": {
                                        "type": "string",
                                        "example": "123 Main St",
                                        "nullable": true
                                    },
                                    "country": {
                                        "type": "string",
                                        "example": "Bangladesh",
                                        "nullable": true
                                    },
                                    "state": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Dhaka",
                                        "nullable": true
                                    },
                                    "zip_code": {
                                        "type": "string",
                                        "example": "1000",
                                        "nullable": true
                                    },
                                    "mobile": {
                                        "type": "string",
                                        "example": "01700000000",
                                        "nullable": true
                                    },
                                    "alternate_number": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "website": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Business registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Business registered successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "access_token": {
                                                    "description": "Passport bearer token — use this for all subsequent requests",
                                                    "type": "string"
                                                },
                                                "token_type": {
                                                    "type": "string",
                                                    "example": "Bearer"
                                                },
                                                "user": {
                                                    "type": "object"
                                                },
                                                "business": {
                                                    "type": "object"
                                                },
                                                "location": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Server error"
                    }
                }
            }
        },
...
        "/v1/business/toggle-active": {
            "patch": {
                "tags": [
                    "Business - Settings"
                ],
                "summary": "Toggle business active state",
                "description": "Toggle the active/inactive state of the authenticated user's business",
                "operationId": "4c66909f2f96f8e778ff4fc65287aff4",
                "responses": {
                    "200": {
                        "description": "Business active state toggled"
                    },
                    "404": {
                        "description": "Business not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}